+++ src/geometry/sphere.rs
@@ -4,12 +4,12 @@ use crate::geometry::material::Material;
 #[derive(Debug, Clone, Copy)]
 pub struct Sphere {
     pub center: Vector,
-    pub radius: f32,
+    pub radius: f64,
     pub material: Material
 }
 
 impl Sphere {
-    pub const fn new(center: Vector, radius: f32, material: Material) -> Sphere {
+    pub const fn new(center: Vector, radius: f64, material: Material) -> Sphere {
         Sphere { center, radius, material }
     }
