+++ src/main.rs
@@ -10,7 +10,7 @@ fn main() {
     let mut image = RgbImage::new(WIDTH, HEIGHT);
 
     for x in 0..WIDTH {
-        utils::print_progress(x as f32 / WIDTH as f32);
+        utils::print_progress(x as f64 / WIDTH as f64);
 
         for y in 0.. HEIGHT {
             image.put_pixel(x, HEIGHT - y - 1, utils::to_rgb(&pixel_color(x, y)));
